+++ docs/screenshots/basic_app.py
@@ -2,6 +2,16 @@ from argparse import Namespace
 
 from complexitty.complexitty import Complexitty
 
-app = Complexitty(Namespace(theme="textual-dark"))
+app = Complexitty(
+    Namespace(
+        colour_map=None,
+        max_iteration=None,
+        multibrot=None,
+        theme="textual-dark",
+        x_position=None,
+        y_position=None,
+        zoom=None,
+    )
+)
 if __name__ == "__main__":
     app.run()