+++ src/complexitty/mandelbrot/__init__.py
@@ -2,10 +2,11 @@
 
 ##############################################################################
 # Local imports.
+from .colouring import colour_maps, get_colour_map
 from .widget import Mandelbrot
 
 ##############################################################################
 # Exports.
-__all__ = ["Mandelbrot"]
+__all__ = ["Mandelbrot", "colour_maps", "get_colour_map"]
 
 ### __init__.py ends here